requirements.txt
agate==1.9.0
agate-dbf==0.2.2
agate-excel==0.4.1
agate-sql==0.7.0
altair==5.4.1
asgiref==3.8.1
attrs==24.2.0
Babel==2.14.0
beautifulsoup4==4.11.1
blinker==1.9.0
cachetools==5.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
csvkit==1.3.0
dbfread==2.0.7
distlib==0.3.9
Django==5.1.4
djangorestframework==3.15.2
et-xmlfile==1.1.0
filelock==3.17.0
Flask==2.1.2
gitdb==4.0.11
GitPython==3.1.43
idna==3.3
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
leather==0.3.4
markdown-it-py==3.0.0
MarkupSafe==2.1.1
mdurl==0.1.2
mmath==1.0.5
mpmath==1.3.0
narwhals==1.14.1
numpy==1.22.3
olefile==0.47
openpyxl==3.1.2
packaging==24.2
pandas==1.4.2
parsedatetime==2.6
Pillow==9.1.0
platformdirs==4.3.6
protobuf==5.28.3
pyarrow==18.0.0
pydeck==0.9.1
Pygments==2.18.0
python-dateutil==2.8.2
python-slugify==8.0.1
pytimeparse==1.1.8
pytz==2022.1
referencing==0.35.1
requests==2.27.1
rich==13.9.4
rpds-py==0.21.0
six==1.16.0
smmap==5.0.1
soupsieve==2.3.2.post1
SQLAlchemy==2.0.23
sqlparse==0.4.2
streamlit==1.40.1
sympy==1.12
tenacity==9.0.0
text-unidecode==1.3
toml==0.10.2
tornado==6.4.1
typing_extensions==4.12.2
urllib3==1.26.9
virtualenv==20.29.1
Werkzeug==2.1.2
xlrd==2.0.1
